pyproject.toml
[project.urls]
homepage = "https://github.com/hu-po/tatbot"
issues = "https://github.com/hu-po/tatbot/issues"
discord = "https://discord.gg/pPAFwndTJd"

[project]
name = "tatbot"
version = "0.5.4"
description="tattoo robot"
authors = [
    { name = "Hugo Ponte", email = "hu.po.xyz@gmail.com" },
]
license = "MIT"
readme = "README.md"
requires-python = "==3.11.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: POSIX :: Linux",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "huggingface-hub==0.34.3",
    "jaxtyping>=0.2.25,<1.0.0",
    "jaxlie>=1.3.4,<2.0.0",
    "mcp==1.11.0",
    "paramiko==3.5.1",
    "pyroki @ git+https://github.com/chungmin99/pyroki.git@main#egg=pyroki",
    "pyyaml==6.0.2",
    "safetensors==0.5.3",
    "tyro==0.9.24",
    "hydra-core~=1.3",
    "pydantic~=2.7",
    "omegaconf~=2.3",
    "hydra-zen",
    "pydantic-numpy",
]

[project.optional-dependencies]
bot = [
    "lerobot[tatbot,smolvla] @ git+https://github.com/hu-po/lerobot.git@main",
    "trossen-arm==1.8.5",
    "evdev==1.9.2",
]
cam = [
    "lerobot[intelrealsense] @ git+https://github.com/hu-po/lerobot.git@main",
    "pupil-apriltags==1.0.4.post11",
    "pyrealsense2==2.55.1.6486",
]
dev = [
    "isort",
    "ruff",
    "pydantic-yaml",
]
gen = [
    "open3d>=0.18.0",
    "potpourri3d==1.3",
]
gpu = [
    "jax[cuda12]>=0.4.0,<0.5.0",
]
img = [
    "opencv-python==4.11.0.86",
    "Pillow>=9.0.0,<11.0.0",
]
viz = [
    "viser==1.0.0",
    "pyliblzfse>=0.4.1",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["bot", "cam", "data", "gen", "mcp", "ops", "utils", "viz"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 110
target-version = "py310"
exclude = ["tests/artifacts/**/*.safetensors", "*_pb2.py", "*_pb2_grpc.py"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "N", "B", "C4", "SIM"]
